\name{boundsdataSED}
\docType{boundsdataSED}
\alias{boundsdataSED}
\title{Data for bounds examples}
\description{
A subsample of the data used in Imai, Tingley, Yamamoto, JRSS-A to illustrate bounds calculation for single experiment design.
}
\usage{data}
\format{A data matrix containing no missing values and is provided only for illustrative purposes.}
\source{Replication data for Imai, Tingley, Yamamoto, JRSS-A.}
\references{
Imai, K., Tingley, D. and Yamamoto, T. (2012) Experimental Designs for Identifying Causal Mechanisms. Journal of the Royal Statistical Society, Series A (Statistics in Society)
}
\keyword{datasets}
