% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.se.R
\name{compute.se}
\alias{compute.se}
\title{Standard error of mediating variable}
\usage{
compute.se(alpha, beta, sigma_alpha, sigma_beta)
}
\arguments{
\item{alpha}{vector of length N}

\item{beta}{vector of length N}
}
\description{
Compute standard error
}
