\name{framing}
\docType{data}
\alias{framing}
\title{Brader, Valentino and Suhay (2008) Framing Experiment Data}
\description{
The \code{framing} data contains 265 rows and 10 columns of data from a framing experiment conducted by Brader, Valentino and Suhay (2008). 
}
\usage{framing}
\format{A data frame containing the following variables:

	\describe{
	\item{immigr}{ a four-point scale measuring subjects' attitudes toward increased immigration.}
	\item{tone}{ 1st treatment; whether the news story is framed positively or negatively.}
	\item{eth}{ 2nd treatment; whether the news story features a Latino or European immigrant.}
	\item{tone_eth}{ product of the two treatment variables. In the original study the authors only find this cell to be significant.}
	\item{emo}{ measure of subjects' anxiety during the experiment. A numeric scale ranging between 3 and 12.}
	\item{p_harm}{ subjects' perceived harm caused by increased immigration. A numeric scale between 2 and 8.}
	\item{ppage}{ subjects' age.}
	\item{ppeducat}{ subjects' highest educational attainments.}
	\item{ppgender}{ subjects' gender.}
	\item{ppincimp}{ subjects' income, measured as a 19-point scale.}
	}

}
\source{

	Brader, T., Valentino, N. and Suhay, E. (2008). What triggers public opposition to immigration? Anxiety, group cues, and immigration threat. American Journal of Political Science 52, 4, 959--978.
	
}
