\name{job}
\docType{data}
\alias{job}
\title{JOB II data}
\description{
Job Search Intervention Study (JOBS
II). JOBS II is a randomized field experiment that investigates the efficacy of a job training
intervention on unemployed workers. The program is designed to not only increase reemployment
among the unemployed but also enhance the mental health of the job seekers. In the JOBS II
field experiment, 1,801 unemployed workers received a pre-screening questionnaire and were then
randomly assigned to treatment and control groups. Those in the treatment group participated
in job-skills workshops. In the workshops, respondents learned job-search skills and coping strategies
for dealing with setbacks in the job-search process. Those in the control condition received
a booklet describing job-search tips. In follow-up interviews, the two key outcome variables were
measured; a continuous measure of depressive symptoms based on the Hopkins Symptom Checklist,
and a binary variable, representing whether the respondent had become employed.
}
\usage{data}
\format{A data matrix containing no missing values and hence is provided only for illustrative purposes and not for inference about program efficacy.}
\source{The complete JOBS II data is available from the data archives at www.icpsr.umich.edu/}
\references{
Vinokur, A. and Schul, Y. (1997). Mastery and inoculation against setbacks as active ingredients
in the jobs intervention for the unemployed. Journal of Consulting and Clinical Psychology 65,
5.
}
\keyword{datasets}
