% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmFit.varyingCovariates.R
\name{lm.varyingCovariates}
\alias{lm.varyingCovariates}
\title{linea model for genomewide expression data with gene-specific covariates}
\usage{
lm.varyingCovariates(Y, X, M)
}
\arguments{
\item{Y}{gene by sample expression matrix (G by N). Required to be normalized and log transformed.}

\item{X}{sample condition labels, assumed to be binary for now.}

\item{M}{data for the mediating variable, a gene by sample matrix (G by N).}
}
