% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmFit.varyingCovariates.voom.R
\name{lm.varyingCovariates.voom}
\alias{lm.varyingCovariates.voom}
\title{linea model for genomewide expression data with gene-specific covariates including voom weights}
\usage{
lm.varyingCovariates.voom(Y, X, M, weights)
}
\arguments{
\item{Y}{gene by sample expression matrix (G by N). Required to be normalized and log transformed.}

\item{X}{sample condition labels, assumed to be binary for now.}

\item{M}{data for the mediating variable, a gene by sample matrix (G by N).}
}
