% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediate.simulate.R
\name{mediate.simulate}
\alias{mediate.simulate}
\title{Simulate data for mediational analysis}
\usage{
mediate.simulate(N_each, G, Y_means, Y_sds, M_means, M_sds)
}
\arguments{
\item{N_each}{number of samples each group}

\item{Y_means, }{Y_sds dependent variable means and standard deviations for each group}

\item{M_means, }{M_sds mediating variable means and standard deviations for each group}
}
\examples{
df.sim <- mediate.simulate(N_each=5, G=50, Y_means = c(0,0), Y_sd=c(1,1), M_means = c(0,1), M_sds = c(1,1))
test.sobel <- mediate.test(Y=df.sim$Y, X=df.sim$X, M=df.sim$M)
}
