% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediate.test.regressing.R
\name{mediate.test.regressing}
\alias{mediate.test.regressing}
\title{An ad hoc approach for testing mediation effect}
\usage{
mediate.test.regressing(Y, X, M, cov = NULL)
}
\arguments{
\item{Y}{gene by sample matrix (G by N) of log2 normalize gene expression.}

\item{X}{vector of length N; sample condition labels, assumed to be binary for now.}

\item{M}{data for the mediating variable, gene by sample matrix (G by N).}
}
\description{
Different from the joint modeling approach, the direct effect is estimated using data after regressing the effect of mediating variable out from the dependent variable. In addition, voom transformation is applied to the expression data for estimating the condition effect.
}
\examples{
#library(medinome)
#library(limma)
#df <- get(load(file="../data/example-kidney-v-heart-human.rda"))
#Y <- df$exprs_pair
#M <- df$methyl_pair
#X <- df$tissue
#cov <- df$RIN
# include covariate
# fit <- mediate.test.regressing(Y=Y, X=X, M=M, cov=NULL)

# not including covariate
# fit <- mediate.test.regressing(Y=Y, X=X, M=M, cov=cov)

#ash_reg <- ash(betahat=fit$d, sebetahat=fit$d_se, lik=lik_normal())

}
