% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediate.test.voom.R
\name{mediate.test.voom}
\alias{mediate.test.voom}
\title{Mediation test for data after voom transformation}
\usage{
mediate.test.voom(Y, X, M)
}
\arguments{
\item{Y}{gene by sample expression matrix (G by N). Required to be normalized and log transformed.}

\item{X}{sample condition labels, assumed to be binary for now.}

\item{M}{data for the mediating variable, a gene by sample matrix (G by N).}
}
\description{
Compute statistics for evaluating mediating effect in genomewide expression data. Apply voom transform to log2 expression data.
}
