% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localFDR.R
\name{localFDR}
\alias{localFDR}
\title{localFDR: A function that fits a Gaussian Mixture model to the mediation coefficients and returns the localFDR estimated from the mixture model}
\usage{
localFDR(
  alpha,
  beta,
  var_alpha,
  var_beta,
  lambda.init = c(0.7, 0.1, 0.1, 0.1),
  kappa.init = 1,
  psi.init = 1,
  psi_int = NULL,
  kappa_int = NULL,
  twostep = FALSE,
  k = 4,
  d1 = NULL,
  d2 = NULL,
  eps = 0.01,
  verbose = TRUE
)
}
\arguments{
\item{alpha}{a vector of estimated alpha coefficients from the first equation of mediation analysis}

\item{beta}{a vector of estimated beta coefficients from the second equation of mediation analysis}

\item{var_alpha}{a vector of estimated variances for alpha coefficients}

\item{var_beta}{a vector of estimated variances for beta coefficients}

\item{lambda.init}{initial values of the proportion of mixture, must sum to 1}

\item{kappa.init}{inital value of kappa, the variance of the prior of alpha under the alternative}

\item{psi.init}{inital value of psi, the variance of the prior of beta under the alternative}

\item{psi_int}{gridSearch interval for psi, to be used in optimize function}

\item{kappa_int}{gridSearch interval for kappa, to be used in optimize function}

\item{twostep}{logical, whether to use two-step MLFDR}

\item{k}{number of mixture components, default is 4. Used in one-step MLFDR.}

\item{d1}{number of non-null components for alpha in two-step MLFDR.}

\item{d2}{number of non-null components for beta in two-step MLFDR.}

\item{eps}{stopping criteria for EM algorithm}

\item{verbose}{logical, whether to print the log-likelihood at each iteration. Default is TRUE.}
}
\value{
A vector of local false discovery rates
}
\description{
localFDR: A function that fits a Gaussian Mixture model to the mediation coefficients and returns the localFDR estimated from the mixture model
}
