% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\docType{package}
\name{mediation}
\alias{mediation}
\alias{mediation-package}
\title{mediation: Parametric and non parametric mediation analysis.}
\description{
mediation: Parametric and non parametric mediation analysis.
}
