% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLFDR.R
\name{MLFDR}
\alias{MLFDR}
\title{MLFDR: Mediation analysis using localFDR}
\usage{
MLFDR(lfdr, size = 0.05)
}
\arguments{
\item{lfdr}{a vector of local false discovery rates, as output from the function localFDR}

\item{size}{Target False Discovery rate, default is 0.05}
}
\value{
A vector of logicals indicating which hypotheses are rejected
}
\description{
MLFDR: Mediation analysis using localFDR
}
